Fix string to cstring memory issue
Any variable that is statically initialized inside a function is put
into stack, which in turn may be overwritten. A pointer to such a
variable inside a function was being returned and used afterwards.
The problem could be fixed using dynamic memory but the issue was
fixed by simply using a built-in function instead.